Package renamed from com.bertramlabs.plugins.karman.s3 to com.bertramlabs.plugins.karman.aws
