+++ src/groovy/com/bertramlabs/plugins/karman/aws/S3Directory.groovy
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
  
-package com.bertramlabs.plugins.karman.s3
+package com.bertramlabs.plugins.karman.aws
 
 import com.amazonaws.services.s3.AmazonS3Client
 import com.amazonaws.services.s3.model.ListObjectsRequest
