+++ src/groovy/com/bertramlabs/plugins/karman/aws/S3StorageProvider.groovy
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.bertramlabs.plugins.karman.s3
+package com.bertramlabs.plugins.karman.aws
 
 import com.amazonaws.auth.BasicAWSCredentials
 import com.amazonaws.regions.Region
