+++ test/unit/com/bertramlabs/plugins/karman/aws/S3CloudFileSpec.groovy
@@ -1,4 +1,4 @@
-package com.bertramlabs.plugins.karman.s3
+package com.bertramlabs.plugins.karman.aws
 
 import com.amazonaws.services.s3.AmazonS3Client
 import com.amazonaws.services.s3.Headers
